refactor(data-dict): split render_readme.py into render/ package
