+++ src/prenatalppkt/scripts/data_dict/render/__init__.py
@@ -1,0 +1,1 @@
+"""Render subpackage: turn comparison.csv into the data-dictionary README."""
